refactor: correct import path for password protection utilities and add JWT cookie management functions for authentication
