+++ ui/src/utils/passwordProtectionUtils.ts
